Cleaned up info and notes
