--- client/src/Components/Accordion/Item/ItemBody/SongInfo/styles.scss
@@ -1,8 +1,0 @@
-@use 'Styles/funcs';
-@use 'Styles/vars';
-
-.song-info {
-  @include funcs.flexCenter();
-  justify-content: flex-start;
-  margin-top: 10px;
-}
