+++ client/src/Components/Accordion/Item/ItemBody/styles.scss
@@ -3,7 +3,7 @@
 
 $height: 60px;
 
-.accordionBody {
+.item-body {
   display: none;
   color: vars.$tertiaryColor;
   font-size: 2.2rem;
@@ -26,16 +26,10 @@ $height: 60px;
     flex-direction: column;
     padding: 5px 10px 30px 10px;
 
-    p {
-      textarea {
-        border: none;
-        background: none;
-        color: whitesmoke;
-        font-size: inherit;
-        width: 340px;
-        height: 100px;
-        resize: none;
-      }
+    section {
+      @include funcs.flexCenter();
+      flex-direction: row;
+      justify-content: flex-start;
     }
   }
 }
