+++ client/src/Components/Accordion/Item/ItemHeader/ItemHeader.scss
@@ -36,7 +36,7 @@ $height: 50px;
         white-space: nowrap;
         text-overflow: ellipsis;
         color: vars.$white;
-        font-size: 1.4rem;
+        font-size: 1.8rem;
         font-weight: 400;
       }
