+++ client/src/Styles/_vars.scss
@@ -9,9 +9,10 @@ $tertiaryColor: #5ae7ff;
 
 // colors for dark theme
 $black: #1d1d1d;
-$gray-light: #959595;
-$gray: #3e3e3e;
 $gray-dark: #2b2b2b;
+$gray: #3e3e3e;
+$gray-light: #959595;
+$gray-extra-light: #d5d5d5;
 $white: #e1e1e1;
 
 // font
